Test: Separate common, admin, and public UI utility functions
