+++ tests/robot-tests/tests/libs/admin-utilities.py
@@ -1,0 +1,23 @@
+from robot.libraries.BuiltIn import BuiltIn
+sl = BuiltIn().get_library_instance('SeleniumLibrary')
+
+def data_csv_number_contains_xpath(num, xpath):
+    try:
+        elem = sl.driver.find_element_by_xpath(f'//*[@id="dataFileUploadForm"]/dl[{num}]')
+    except:
+        raise AssertionError(f'Cannot find data file number "{num}"')
+    try:
+        elem.find_element_by_xpath(xpath)
+    except:
+        raise AssertionError(f'Cannot find data file number "{num} with xpath {xpath}')
+
+def data_file_number_contains_xpath(num, xpath):
+    try:
+        elem = sl.driver.find_element_by_xpath(f'//*[@id="fileUploadForm"]/dl[{num}]')
+    except:
+        raise AssertionError(f'Cannot find data file number "{num}"')
+    try:
+        elem.find_element_by_xpath(xpath)
+    except:
+        raise AssertionError(f'Cannot find data file number "{num} with xpath {xpath}')
+
